Fix AttributeError: 'RelationshipPropertiesExtractor' object has no attribute 'assay_enzyme_relationship'. Did you mean: 'assay_gene_relationship'?

+++ chemgraphbuilder/node_data_processor.py
@@ -18,12 +18,16 @@ Example Usage:
     >>> processor.preprocess_genes()
     >>> processor.preprocess_compounds()
 """
+
 import glob
 import pandas as pd
 import logging
 
 # Set up logging configuration
-logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
+logging.basicConfig(
+    level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s"
+)
+
 
 class NodeDataProcessor:
     """
@@ -52,43 +56,60 @@ class NodeDataProcessor:
         """
         self.data_dir = data_dir
 
-
     def preprocess_assays(self):
         """
         Processes the assay data by renaming columns and saving the modified data back to disk.
         This method also handles visualization of assay data distributions if necessary.
         """
-        df = pd.read_csv(f'{self.data_dir}/Nodes/Assay_Properties.csv')
-        df.rename(columns={"AID": "AssayID", "Assay Type": "AssayType",
-                           "Activity Name": "AssayActivityName", "SourceID": "AssaySourceID",
-                           "SourceName": "AssaySourceName", "Name": "AssayName",
-                           "Description": "AssayDescription"}, inplace=True)
-        df.to_csv(f'{self.data_dir}/Nodes/Assay_Properties_Processed.csv', index=False)
-
+        df = pd.read_csv(f"{self.data_dir}/Nodes/Assay_Properties.csv")
+        df.rename(
+            columns={
+                "AID": "AssayID",
+                "Assay Type": "AssayType",
+                "Activity Name": "AssayActivityName",
+                "SourceID": "AssaySourceID",
+                "SourceName": "AssaySourceName",
+                "Name": "AssayName",
+                "Description": "AssayDescription",
+            },
+            inplace=True,
+        )
+        df.to_csv(f"{self.data_dir}/Nodes/Assay_Properties_Processed.csv", index=False)
 
     def preprocess_proteins(self):
         """
         Processes the protein data by renaming columns and saving the processed data.
         This method simplifies access to protein data for downstream analysis.
         """
-        df = pd.read_csv(f'{self.data_dir}/Nodes/Protein_Properties.csv')
-        df.rename(columns={"RefSeq Accession": "ProteinRefSeqAccession",
-                           "Description": "ProteinDescription"}, inplace=True)
-        df.to_csv(f'{self.data_dir}/Nodes/Protein_Properties_Processed.csv', index=False)
-
+        df = pd.read_csv(f"{self.data_dir}/Nodes/Protein_Properties.csv")
+        df.rename(
+            columns={
+                "RefSeq Accession": "ProteinRefSeqAccession",
+                "Description": "ProteinDescription",
+            },
+            inplace=True,
+        )
+        df.to_csv(
+            f"{self.data_dir}/Nodes/Protein_Properties_Processed.csv", index=False
+        )
 
     def preprocess_genes(self):
         """
         Processes gene data by renaming columns and changing data types for specific fields.
         The processed data is saved for further use in gene-related analyses.
         """
-        df = pd.read_csv(f'{self.data_dir}/Nodes/Gene_Properties.csv')
-        df.rename(columns={"Symbol": "GeneSymbol", "Taxonomy ID": "TaxonomyID",
-                           "Synonyms": "GeneSynonyms"}, inplace=True)
-        df['GeneID'] = df['GeneID'].astype('Int64')
-        df['TaxonomyID'] = df['TaxonomyID'].astype('Int64')
-        df.to_csv(f'{self.data_dir}/Nodes/Gene_Properties_Processed.csv', index=False)
-
+        df = pd.read_csv(f"{self.data_dir}/Nodes/Gene_Properties.csv")
+        df.rename(
+            columns={
+                "Symbol": "GeneSymbol",
+                "Taxonomy ID": "TaxonomyID",
+                "Synonyms": "GeneSynonyms",
+            },
+            inplace=True,
+        )
+        df["GeneID"] = df["GeneID"].astype("Int64")
+        df["TaxonomyID"] = df["TaxonomyID"].astype("Int64")
+        df.to_csv(f"{self.data_dir}/Nodes/Gene_Properties_Processed.csv", index=False)
 
     def preprocess_compounds(self):
         """
@@ -96,14 +117,14 @@ class NodeDataProcessor:
         renames columns for uniformity, and saves the consolidated data. This method
         facilitates easier management and analysis of compound data.
         """
-        path = f'{self.data_dir}/Nodes/Compound_Properties'
+        path = f"{self.data_dir}/Nodes/Compound_Properties"
         all_csv_files = glob.glob(path + "/*.csv")
         first_file = True
-        output_file = f'Data/Nodes/Compound_Properties.csv'
+        output_file = f"Data/Nodes/Compound_Properties.csv"
 
-        with open(output_file, 'w', newline='', encoding='utf-8') as f_out:
+        with open(output_file, "w", newline="", encoding="utf-8") as f_out:
             for file in all_csv_files:
-                with open(file, 'r', newline='', encoding='utf-8') as f_in:
+                with open(file, "r", newline="", encoding="utf-8") as f_in:
                     header = f_in.readline()
                     if first_file:
                         f_out.write(header)
